+++ src/extensions/core/uploadAudio.ts
@@ -2,7 +2,7 @@
 import type { IWidget } from '@comfyorg/litegraph'
 import type { IStringWidget } from '@comfyorg/litegraph/dist/types/widgets'
 
-import { useNodeFileInput } from '@/composables/useNodeFileInput'
+import { useNodeFileInput } from '@/composables/node/useNodeFileInput'
 import type { DOMWidget } from '@/scripts/domWidget'
 import { useToastStore } from '@/stores/toastStore'
 import { ComfyNodeDef } from '@/types/apiTypes'
